requirements.txt
pyautogui==0.9.53
Pillow==8.3.2
python-imagesearch==1.2.0